refactor: add omnyssh-core crate, move utils/platform and utils/scroll
