+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["crates/omnyssh"]
+members = ["crates/omnyssh-core", "crates/omnyssh"]
 
 [workspace.package]
 version = "1.1.0"
