+++ crates/omnyssh-core/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "omnyssh-core"
+version.workspace = true
+edition.workspace = true
+description = "SSH engine, configs, metrics, and updater powering OmnySSH frontends"
+license.workspace = true
+repository.workspace = true
+keywords = ["ssh", "devops"]
+categories = ["network-programming"]
+
+[dependencies]
+# Cross-platform paths (P-CROSS-01)
+dirs = { workspace = true }
+
+# Terminal emulator — vt100 screen model (PTY is provided by russh channels)
+vt100 = { package = "vt100-omnyssh", version = "0.15" }
