+++ crates/omnyssh-core/src/utils/mod.rs
@@ -1,3 +1,3 @@
 /// Cross-platform utility helpers.
-pub mod mouse;
-pub mod paste;
+pub mod platform;
+pub mod scroll;
