+++ crates/omnyssh-core/src/utils/platform.rs
