+++ crates/omnyssh-core/src/utils/scroll.rs
