+++ crates/omnyssh/Cargo.toml
@@ -17,6 +17,8 @@ name = "omny"
 path = "src/main.rs"
 
 [dependencies]
+omnyssh-core = { path = "../omnyssh-core", version = "1.1.0" }
+
 # TUI
 ratatui = "0.29"
 crossterm = "0.28"
