+++ crates/omnyssh/src/app/terminal.rs
@@ -114,11 +114,11 @@ impl App {
         };
         // Inspect the foreground app under a brief parser lock, then release it.
         let action = match tab.parser.lock() {
-            Ok(parser) => crate::utils::scroll::resolve_scroll(delta, parser.screen()),
+            Ok(parser) => omnyssh_core::utils::scroll::resolve_scroll(delta, parser.screen()),
             Err(_) => return,
         };
         match action {
-            crate::utils::scroll::ScrollAction::Scrollback(d) => {
+            omnyssh_core::utils::scroll::ScrollAction::Scrollback(d) => {
                 if d > 0 {
                     // Cap at the vt100 scrollback capacity (1000 lines, see pty.rs).
                     tab.scroll_offset = tab.scroll_offset.saturating_add(d as usize).min(1000);
@@ -126,7 +126,7 @@ impl App {
                     tab.scroll_offset = tab.scroll_offset.saturating_sub((-d) as usize);
                 }
             }
-            crate::utils::scroll::ScrollAction::Forward(bytes) => {
+            omnyssh_core::utils::scroll::ScrollAction::Forward(bytes) => {
                 let id = tab.session_id;
                 if let Some(mgr) = &mut self.pty_manager {
                     if let Err(e) = mgr.write(id, &bytes) {
