+++ crates/omnyssh/src/config/app_config.rs
@@ -210,7 +210,7 @@ impl Default for KeybindingsConfig {
 /// # Errors
 /// Returns an error only if the file exists but cannot be read or parsed.
 pub fn load_app_config(path: Option<&std::path::Path>) -> anyhow::Result<AppConfig> {
-    use crate::utils::platform;
+    use omnyssh_core::utils::platform;
 
     let config_path = match path {
         Some(p) => p.to_path_buf(),
@@ -239,7 +239,7 @@ pub fn load_app_config(path: Option<&std::path::Path>) -> anyhow::Result<AppConf
 /// # Errors
 /// Returns an error if the config file cannot be read, parsed, or written.
 fn persist_config<F: FnOnce(&mut AppConfig)>(mutator: F) -> anyhow::Result<()> {
-    use crate::utils::platform;
+    use omnyssh_core::utils::platform;
 
     let config_path = match platform::app_config_path() {
         Some(p) => p,
